Created full playing zone, player movement, physics interactions
This will be the last commit for a while as I am shealving this project and Godot. Godot has just had its big 4.2 release finished, but Vulcan 4 (3D graphics) has not been fully implemented yet. In this time of transition I am content with taking my 2D knowledge and moving to Unreal and Unity for more practice in the 3D space.